+++ packages/runtime/src/builder/index.ts
@@ -1,1 +1,2 @@
 export * from './serialization'
+export * from './unstructured-introspection'