+++ packages/runtime/src/next/api-handler/index.ts
@@ -8,15 +8,14 @@ import { Makeswift } from '../client'
 import elementTree, { ElementTreeResponse } from './handlers/element-tree'
 import fonts, { Font, FontsResponse, GetFonts } from './handlers/fonts'
 import manifest, { Manifest, ManifestResponse } from './handlers/manifest'
-import proxyPreviewMode, { ProxyPreviewModeResponse } from './handlers/proxy-preview-mode'
-import proxyDraftMode, { ProxyDraftModeResponse } from './handlers/proxy-draft-mode'
-import draftMode, { type DraftModeResponse } from './handlers/draft-mode'
-import previewMode, { type PreviewModeResponse } from './handlers/preview-mode'
+import redirectDraft, { type RedirectDraftResponse } from './handlers/redirect-draft'
+import redirectPreview, { type RedirectPreviewResponse } from './handlers/redirect-preview'
+import clearDraft, { type ClearDraftResponse } from './handlers/clear-draft'
 import { revalidate, RevalidationResponse } from './handlers/revalidate'
 import translatableData, { TranslatableDataResponse } from './handlers/translatable-data'
 import mergeTranslatedData, { TranslatedDataResponse } from './handlers/merge-translated-data'
 import webhook from './handlers/webhook'
-import { WebhookResponseBody } from './handlers/webhook/types'
+import { OnPublish, WebhookResponseBody } from './handlers/webhook/types'
 import { ReactRuntime } from '../../react'
 import { P, match } from 'ts-pattern'
 import {
@@ -29,10 +28,13 @@ export type { Manifest, Font }
 
 type Context = { params: { [key: string]: string | string[] } }
 
+type Events = { onPublish: OnPublish }
+
 type MakeswiftApiHandlerConfig = {
   appOrigin?: string
   apiOrigin?: string
   getFonts?: GetFonts
+  events?: Events
   runtime: ReactRuntime
 }
 
@@ -41,10 +43,9 @@ type NotFoundError = { message: string }
 export type MakeswiftApiHandlerResponse =
   | ManifestResponse
   | RevalidationResponse
-  | ProxyPreviewModeResponse
-  | ProxyDraftModeResponse
-  | DraftModeResponse
-  | PreviewModeResponse
+  | RedirectDraftResponse
+  | RedirectPreviewResponse
+  | ClearDraftResponse
   | FontsResponse
   | ElementTreeResponse
   | TranslatableDataResponse
@@ -69,6 +70,7 @@ export function MakeswiftApiHandler(
     appOrigin = 'https://app.makeswift.com',
     apiOrigin = 'https://api.makeswift.com',
     getFonts,
+    events,
     runtime,
   }: MakeswiftApiHandlerConfig,
 ): (...args: MakeswiftApiHandlerArgs) => Promise<NextResponse<MakeswiftApiHandlerResponse> | void> {
@@ -174,31 +176,24 @@ export function MakeswiftApiHandler(
         .exhaustive()
     }
 
-    if (matches('/proxy-preview-mode')) {
+    if (matches('/draft')) {
       return match(args)
-        .with(routeHandlerPattern, args => proxyPreviewMode(...args, { apiKey }))
-        .with(apiRoutePattern, args => proxyPreviewMode(...args, { apiKey }))
+        .with(routeHandlerPattern, args => redirectDraft(...args, { apiKey }))
+        .with(apiRoutePattern, args => redirectDraft(...args, { apiKey }))
         .exhaustive()
     }
 
-    if (matches('/proxy-draft-mode')) {
+    if (matches('/preview')) {
       return match(args)
-        .with(routeHandlerPattern, args => proxyDraftMode(...args, { apiKey }))
-        .with(apiRoutePattern, args => proxyDraftMode(...args, { apiKey }))
+        .with(routeHandlerPattern, args => redirectPreview(...args, { apiKey }))
+        .with(apiRoutePattern, args => redirectPreview(...args, { apiKey }))
         .exhaustive()
     }
 
-    if (matches('/draft-mode')) {
+    if (matches('/clear-draft')) {
       return match(args)
-        .with(routeHandlerPattern, args => draftMode(...args, { apiKey }))
-        .with(apiRoutePattern, args => draftMode(...args, { apiKey }))
-        .exhaustive()
-    }
-
-    if (matches('/preview-mode')) {
-      return match(args)
-        .with(routeHandlerPattern, args => previewMode(...args, { apiKey }))
-        .with(apiRoutePattern, args => previewMode(...args, { apiKey }))
+        .with(routeHandlerPattern, args => clearDraft(...args, { apiKey }))
+        .with(apiRoutePattern, args => clearDraft(...args, { apiKey }))
         .exhaustive()
     }
 
@@ -232,8 +227,8 @@ export function MakeswiftApiHandler(
 
     if (matches('/webhook')) {
       return match(args)
-        .with(routeHandlerPattern, args => webhook(...args, { apiKey }))
-        .with(apiRoutePattern, args => webhook(...args, { apiKey }))
+        .with(routeHandlerPattern, args => webhook(...args, { apiKey, events }))
+        .with(apiRoutePattern, args => webhook(...args, { apiKey, events }))
         .exhaustive()
     }