+++ packages/runtime/src/next/cache.ts
@@ -1,0 +1,1 @@
+export const MAKESWIFT_CACHE_TAG = '@@makeswift'