+++ packages/runtime/src/runtimes/react/components/RuntimeProvider.tsx
@@ -7,6 +7,9 @@ import { ReactRuntimeContext } from '../hooks/use-react-runtime'
 import { ReactRuntime } from '../react-runtime'
 import { MakeswiftHostApiClientProvider } from '../host-api-client'
 import { MakeswiftSiteVersion } from '../../../api/site-version'
+import { DraftSwitcher } from './draft-switcher/draft-switcher'
+import { useBuilderHandshake } from './hooks/use-builder-handshake'
+import { useBuilderConnectionPing } from './hooks/use-builder-connection-ping'
 
 const LiveProvider = lazy(() => import('./LiveProvider'))
 const PreviewProvider = lazy(() => import('./PreviewProvider'))
@@ -15,6 +18,7 @@ export function ReactRuntimeProvider({
   children,
   runtime,
   previewMode,
+  appOrigin = 'https://app.makeswift.com',
   apiOrigin = 'https://api.makeswift.com',
   locale = undefined,
 }: {
@@ -22,6 +26,7 @@ export function ReactRuntimeProvider({
   runtime: ReactRuntime
   previewMode: boolean
   apiOrigin?: string
+  appOrigin?: string
   locale?: string
 }) {
   const client = useMemo(
@@ -36,10 +41,16 @@ export function ReactRuntimeProvider({
 
   const StoreProvider = previewMode ? PreviewProvider : LiveProvider
 
+  useBuilderHandshake({ appOrigin })
+  useBuilderConnectionPing({ appOrigin })
+
   return (
     <ReactRuntimeContext.Provider value={runtime}>
       <MakeswiftHostApiClientProvider client={client}>
-        <StoreProvider>{children}</StoreProvider>
+        <StoreProvider>
+          {children}
+          <DraftSwitcher isDraft={previewMode} />
+        </StoreProvider>
       </MakeswiftHostApiClientProvider>
     </ReactRuntimeContext.Provider>
   )