+++ packages/runtime/src/runtimes/react/hooks/use-register-document.ts
@@ -2,7 +2,9 @@ import { useEffect } from 'react'
 import { type Document } from '../../../state/react-page'
 import { useDispatch } from './use-dispatch'
 import { useIsInBuilder } from './use-is-in-builder'
-import { registerBuilderDocumentsEffect, registerDocumentsEffect } from '../../../state/actions'
+import { registerBuilderDocumentsEffect, registerDocument, registerDocumentsEffect } from '../../../state/actions'
+import { isServer } from '../../../utils/is-server'
+import { useIsomorphicLayoutEffect } from '../../../components/hooks/useIsomorphicLayoutEffect'
 
 /**
  * @param document Document to register
@@ -11,7 +13,17 @@ export function useRegisterDocument(document: Document): void {
   const isInBuilder = useIsInBuilder()
   const dispatch = useDispatch()
 
-  useEffect(() => dispatch(registerDocumentsEffect([document])), [document])
+  if (isServer()) {
+    dispatch(registerDocument(document))
+  }
+
+  /*
+    Layout effect is to ensure that the document registration happens prior to the
+    attempted creation/registration of prop controllers in child components.
+  */
+  useIsomorphicLayoutEffect(() => {
+    return dispatch(registerDocumentsEffect([document]))
+  }, [dispatch, document])
 
   // TODO: Decide whether to do this via middleware or via explicit action (like
   // what we're doing below)