+++ packages/prop-controllers/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @makeswift/prop-controllers
 
+## 0.4.3
+
+### Patch Changes
+
+- Updated dependencies [f8720ba]
+  - @makeswift/controls@0.1.10
+
+## 0.4.2
+
+### Patch Changes
+
+- f33c4d7: Add support for removing resources while copying control/prop-controller data with a replacement context
+- Updated dependencies [f33c4d7]
+  - @makeswift/controls@0.1.9
+
 ## 0.4.1
 
 ### Patch Changes